Fixed issues with dragging calculator
