+++ src/components/distance-calculator/CalculatorTab.tsx
@@ -1,57 +1,72 @@
-import {
-  Box, IconButton, MenuItem, Stack, TextField, Typography,
-} from '@mui/material';
+import { useSortable } from '@dnd-kit/sortable';
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { CSS } from '@dnd-kit/utilities';
+import { Box, IconButton, Stack } from '@mui/material';
 import { EntityId } from '@reduxjs/toolkit';
 import { shallowEqual } from 'react-redux';
 
+import DistanceForm from './DistanceForm';
 import XIcon from '../../assets/x-icon.png';
 import { useAppDispatch } from '../../store';
 import { useDistanceCalculatorSelector } from '../../store/distance-calculator';
 import {
   calculatorTabsSelectors, deleteTab, switchTab, tabCalculatorSelector, updateTab,
 } from '../../store/distance-calculator/calculatorsSlice';
-import { PENALTIES } from '../../utils/gameData';
-import { calcDistance, calcTravelTime } from '../../utils/math';
-import { formatSeconds } from '../../utils/utilities';
-import CoordinateInput from '../common/CoordinateInput';
 import CustomIcon from '../common/CustomIcon';
-import SimpleNumberField from '../common/SimpleNumberField';
 import TypographyTextField from '../common/TypographyTextField';
 
 type CalculatorTabProps = {
   tabId: EntityId,
+  // False positive?
+  // eslint-disable-next-line react/no-unused-prop-types
+  asOverlay?: boolean,
 };
 
-function CalculatorTabButton({ tabId }: CalculatorTabProps) {
-  const dispatch = useAppDispatch();
+function CalculatorTabButton({ tabId, asOverlay = false }: CalculatorTabProps) {
+  const {
+    attributes,
+    listeners,
+    setNodeRef,
+    transform,
+    transition,
+  } = useSortable({ id: tabId });
 
+  const dispatch = useAppDispatch();
   const active = useDistanceCalculatorSelector((state) => {
-    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId);
-    if (!tab) return false;
-    const calculator = tabCalculatorSelector(state.calculators, tab.id);
-    if (!calculator) return false;
+    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId)!;
+    const calculator = tabCalculatorSelector(state.calculators, tab.id)!;
     return calculator.currentTab === tabId;
   });
   const calculatorId = useDistanceCalculatorSelector((state) => {
-    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId);
-    if (!tab) return null;
-    return tabCalculatorSelector(state.calculators, tab.id)?.id ?? null;
+    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId)!;
+    return tabCalculatorSelector(state.calculators, tab.id)!.id;
   }, shallowEqual);
   const tabName = useDistanceCalculatorSelector((state) => {
-    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId);
-    if (!tab) return null;
+    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId)!;
     return tab.name ?? '';
   });
+  const dragging = useDistanceCalculatorSelector(
+    (state) => (state.draggingTab === tabId) && !asOverlay,
+  );
 
-  if (tabName === null || calculatorId === null) return null;
   return (
     <Box
+      ref={setNodeRef}
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      {...attributes}
+      // eslint-disable-next-line react/jsx-props-no-spreading
+      {...listeners}
       className="clickable center-content"
       onClick={() => dispatch(switchTab({ calculatorId, tabId }))}
       sx={{
         padding: '2px 10px',
-        borderRight: '1px solid black',
-        borderBottom: `1px solid ${active ? 'transparent' : 'black'}`,
+        borderRight: dragging ? undefined : '1px solid black',
+        borderBottom: dragging ? undefined : `1px solid ${active ? 'transparent' : 'black'}`,
+        backgroundColor: 'white',
+        transform: CSS.Transform.toString(transform),
+        transition,
+        opacity: asOverlay ? 0.5 : undefined,
+        pointerEvents: asOverlay ? 'none' : undefined,
       }}
     >
       <Stack direction="row" alignItems="center" spacing={1}>
@@ -88,107 +103,9 @@ function CalculatorTabButton({ tabId }: CalculatorTabProps) {
 }
 
 function CalculatorTab({ tabId }: CalculatorTabProps) {
-  const dispatch = useAppDispatch();
-  const active = useDistanceCalculatorSelector((state) => {
-    const tab = calculatorTabsSelectors.selectById(state.tabs, tabId);
-    if (!tab) return false;
-    const calculator = tabCalculatorSelector(state.calculators, tab.id);
-    if (!calculator) return false;
-    return calculator.currentTab === tab.id;
-  });
-  const tab = useDistanceCalculatorSelector(
-    (state) => calculatorTabsSelectors.selectById(state.tabs, tabId) ?? null,
-    shallowEqual,
-  );
-  if (!tab || !active) return null;
-
-  const distance = calcDistance(tab.startingPoint, tab.endingPoint, tab.penalty);
-  const time = calcTravelTime(distance, tab.speed);
-
   return (
     <Box sx={{ padding: '20px' }}>
-      <Stack spacing={3}>
-        <Stack alignItems="center" spacing={1}>
-          <Typography>Starting point</Typography>
-          <CoordinateInput
-            value={tab.startingPoint}
-            onChange={(p) => dispatch(updateTab({ tabId, update: { startingPoint: p } }))}
-          />
-        </Stack>
-        <Stack alignItems="center" spacing={1}>
-          <Typography>Ending point</Typography>
-          <CoordinateInput
-            value={tab.endingPoint}
-            onChange={(p) => dispatch(updateTab({ tabId, update: { endingPoint: p } }))}
-          />
-        </Stack>
-        <TextField
-          select
-          label="Mission Penalty"
-          value={tab.penalty}
-          onChange={(e) => dispatch(
-            updateTab({ tabId, update: { penalty: Number(e.target.value) } }),
-          )}
-          SelectProps={{
-            renderValue: (v) => {
-              const penalty = Object.values(PENALTIES).find((p) => p.value === v);
-              if (!penalty) return null;
-              return (
-                <Stack direction="row" spacing={1}>
-                  <CustomIcon src={penalty.iconSrc} tintColor={penalty.iconColor} />
-                  <Typography>
-                    (+
-                    { penalty.value }
-                    )
-                    { ' ' }
-                    { penalty.name }
-                  </Typography>
-                </Stack>
-              );
-            },
-          }}
-        >
-          { Object.values(PENALTIES).map((penalty) => (
-            <MenuItem key={penalty.name} value={penalty.value}>
-              <Stack>
-                <Stack direction="row" spacing={1}>
-                  <CustomIcon src={penalty.iconSrc} tintColor={penalty.iconColor} />
-                  <Typography variant="subtitle1">
-                    (+
-                    { penalty.value }
-                    )
-                    { ' ' }
-                    { penalty.name }
-                  </Typography>
-                </Stack>
-                <Typography>{ penalty.description }</Typography>
-              </Stack>
-            </MenuItem>
-          )) }
-
-        </TextField>
-        <SimpleNumberField
-          value={tab.speed}
-          onChange={(v) => dispatch(updateTab({ tabId, update: { speed: v } }))}
-          minValue={0}
-          maxValue={150}
-          textFieldProps={{
-            label: 'Warlord Speed',
-          }}
-        />
-        <Stack>
-          <Typography color={!Number.isFinite(distance) ? 'gray' : undefined}>
-            Distance:
-            { ' ' }
-            { !Number.isFinite(distance) ? '∞' : distance }
-          </Typography>
-          <Typography color={!Number.isFinite(time) ? 'gray' : undefined}>
-            Time:
-            { ' ' }
-            { formatSeconds(time) }
-          </Typography>
-        </Stack>
-      </Stack>
+      <DistanceForm tabId={tabId} />
     </Box>
   );
 }
